Add Required changes and fixed linter errors
