+++ src/components/Categories.jsx
@@ -1,0 +1,11 @@
+const Categories = () => {
+  const result = (
+    <div className="CategoriesComponent">
+      <h2 className="notFound"> Under development </h2>
+    </div>
+  );
+
+  return result;
+};
+
+export default Categories;
