+++ src/styles/Css/App.css
@@ -17,11 +17,6 @@ body {
   white-space: nowrap;
 }
 
-body img {
-  max-width: 20px;
-  max-height: 17px;
-}
-
 body h2,
 body span,
 body h3 {
@@ -59,6 +54,10 @@ button:focus {
 .bookstoreTitle {
   color: #1ba8c5;
   font-weight: 700;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 .fullWebsite {
@@ -70,7 +69,7 @@ button:focus {
 }
 
 .App,
-.Books {
+.BooksContainer {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -187,12 +186,12 @@ footer a {
 
 .BookDetails,
 .BookChapter,
-.MainContainer {
+.Books {
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: stretch;
-  gap: 1.5rem;
+  gap: 0.5rem;
 }
 
 .BookDetails {
@@ -281,11 +280,16 @@ nav {
   display: flex;
   flex-direction: row;
   align-items: center !important;
+  justify-content: center;
   gap: 1rem;
 }
 
 .navBar div span {
   align-self: center !important;
+}
+
+.accountPic {
+  max-width: 17px;
 }
 
 .HomeComponent {
@@ -341,6 +345,10 @@ nav {
     font-size: 18px;
   }
 
+  .accountPic {
+    max-width: 25px;
+  }
+
   .navBar {
     padding: 0.2rem 1.5rem;
   }
@@ -353,6 +361,10 @@ nav {
 @media (min-width: 999px) {
   body {
     font-size: 25px;
+  }
+
+  .accountPic {
+    max-width: 30px;
   }
 
   .headSpan {
